Remove handler and migrate tests for metadata.
